asdf-backend/requirements.txt
fastapi==0.112.1
pydantic==2.8.2
python-multipart==0.0.9
uvicorn==0.30.6
openai==1.42.0
python-dotenv==1.0.1